+++ tests/integration/api/test_scans.py
@@ -82,7 +82,8 @@ class TestScansApi(BaseTest):
             ScanLaunchRequest()
         )
         scan_details = client.scans_api.details(scan_id)
-        assert scan_details.info.status in [Scan.STATUS_PENDING, Scan.STATUS_INITIALIZING, Scan.STATUS_RUNNING], u'Scan is in launched state.'
+        assert scan_details.info.status in [Scan.STATUS_PENDING, Scan.STATUS_INITIALIZING, Scan.STATUS_RUNNING], \
+            u'Scan is in launched state.'
 
         scan_details = self.wait_until(lambda: client.scans_api.details(scan_id),
                                        lambda details: details.info.status in [